Remove unnecessary f's from stdlib cmath functions
powf() is a C function but std::pow() is the thin C++ wrapper that
delegates to pow(), powf(), or powl() dependent on argument type.
Most of these are already acting on floats, and those that aren't have
been fixed up to do so.

+++ src/throttle_display.cpp
@@ -87,8 +87,8 @@ void ThrottleDisplay::renderThrottle(const Bounds &bounds) {
 
     // Draw throttle plate
     const float throttleAngle = (float)m_engine->getThrottlePlateAngle();
-    const float cos_theta = std::cosf(throttleAngle);
-    const float sin_theta = std::sinf(throttleAngle);
+    const float cos_theta = std::cos(throttleAngle);
+    const float sin_theta = std::sin(throttleAngle);
 
     params.y0 = origin.y - sin_theta * plateWidth / 2.0f;
     params.x0 = origin.x - cos_theta * plateWidth / 2.0f;
